Fix #86, split configuration
Use the "generate_config_includefile" for all headers that can come from
alternate source (e.g. generated or mission-supplied).
