+++ config/default_lc_tbldefs.h
