+++ docs/dox_src/lc-common.doxyfile.in
@@ -5,5 +5,6 @@
 # Include front material followed by everything in fsw
 INPUT      += @lc_MISSION_DIR@/docs/dox_src/cfs_lc.dox
 INPUT      += @lc_MISSION_DIR@/fsw
+INPUT      += @lc_MISSION_DIR@/config
 
 IMAGE_PATH += @lc_MISSION_DIR@/docs/dox_src
