+++ fsw/src/lc_action.c
@@ -28,8 +28,9 @@
 #include "lc_app.h"
 #include "lc_action.h"
 #include "lc_msgids.h"
-#include "lc_events.h"
+#include "lc_eventids.h"
 #include "lc_custom.h"
+#include "lc_platform_cfg.h"
 
 #include <string.h>
