+++ fsw/src/lc_action.h
@@ -29,6 +29,7 @@
  * Includes
  *************************************************************************/
 #include "cfe.h"
+#include "lc_tbl.h"
 
 /************************************************************************
  * Macro Definitions
