+++ fsw/src/lc_custom.c
@@ -28,9 +28,10 @@
 *************************************************************************/
 #include "lc_custom.h"
 #include "lc_tbldefs.h"
-#include "lc_events.h"
+#include "lc_eventids.h"
 #include "lc_mission_cfg.h"
 #include "lc_msg.h"
+#include "lc_msgids.h"
 
 /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 /*                                                                 */
