+++ fsw/src/lc_utils.c
@@ -28,7 +28,7 @@
 #include "lc_app.h"
 #include "lc_cmds.h"
 #include "lc_msgids.h"
-#include "lc_events.h"
+#include "lc_eventids.h"
 #include "lc_version.h"
 #include "lc_action.h"
 #include "lc_watch.h"
