+++ unit-test/lc_app_tests.c
@@ -28,13 +28,15 @@
 
 #include "lc_app.h"
 #include "lc_msg.h"
+#include "lc_tbldefs.h"
 #include "lc_msgdefs.h"
 #include "lc_msgids.h"
-#include "lc_events.h"
+#include "lc_eventids.h"
 #include "lc_version.h"
 #include "lc_utils.h"
 #include "lc_test_utils.h"
 #include "lc_dispatch.h"
+#include "lc_platform_cfg.h"
 
 #include "cfe.h"
