+++ unit-test/lc_utils_tests.c
@@ -21,9 +21,10 @@
 #include "lc_utils.h"
 #include "lc_app.h"
 #include "lc_msg.h"
+#include "lc_tbldefs.h"
 #include "lc_msgdefs.h"
 #include "lc_msgids.h"
-#include "lc_events.h"
+#include "lc_eventids.h"
 #include "lc_version.h"
 #include "lc_test_utils.h"
